backend/ml-service/requirements.txt
fastapi
uvicorn
paho-mqtt
psycopg2-binary
python-dotenv